chore: adding stress test
